refactor: extract matrix UI controller from webview/tec1g/index.ts
Move all matrix keyboard state, DOM references, functions, and event
handlers into a new webview/tec1g/matrix-ui.ts module behind a
MatrixUiController interface.

index.ts interacts with the matrix subsystem through 5 call sites:
- matrixUi.init() at startup
- matrixUi.applyMatrixRows() in applyUpdate
- matrixUi.applyCapsLock() in applyUpdate
- matrixUi.applyMatrixMode() in applyUpdate
- matrixUi.handleKeyEvent() in keydown/keyup handlers

Closes #120
